pyproject.toml
[project]
name = "apex-ledger"
version = "0.1.0"
description = "Agentic Event Store and Enterprise Audit Infrastructure"
requires-python = ">=3.12"

dependencies = [
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "anthropic>=0.40.0",
    "openai>=2.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apex_ledger"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
